requirements.txt
asgiref==3.6.0
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.6
django-heroku==0.3.1
djangorestframework==3.14.0
filelock==3.9.0
gunicorn==20.1.0
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
virtualenv==20.19.0
whitenoise==6.4.0
